requirements.txt
acres==0.5.0
aiofiles==24.1.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
filelock==3.19.1
frontend==0.0.3
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httplib2==0.31.0
huggingface-hub==0.35.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pillow==11.3.0
pluggy==1.6.0
prov==2.1.1
puremagic==1.30
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-pptx==1.0.2
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==7.2.1
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
traits==7.0.2
transformers==4.57.0
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
xlsxwriter==3.2.9
